+++ template-sample-app/template/src/data/models/storeModel.js
